Adds the housekeeping MCU with CAN bootloader and A/B boot, the FreeRTOS modem on Core 1 with the AMP driver, the ACM/PUS payload manager, and the SCPI server, Qt ground station, Rust CLI, HIL suite, performance report and FDIR chain; everything is verified on the host, and hardware testing is still pending.

+++ linux/drivers/spec_tap/spec_tap_drv.c
@@ -185,7 +185,7 @@ static int satlink_st_probe(struct platform_device *pdev)
 
 	version_reg = readl(st->regs + ST_REG_VERSION);
 	dev_info(dev, "spec_tap v%u.%u ready, DMA %s\n", version_reg >> ST_VERSION_MAJOR_SHIFT,
-		 version_reg & ST_VERSION_MINOR_MASK,
+		 (u32)(version_reg & ST_VERSION_MINOR_MASK),
 		 st->dma_s2mm ? "attached" : "not yet attached");
 
 	return 0;
